feat(cli): edgeplane task mesh — CLI mirror of the mesh-task MCP tools (#111)
EdgePlane's design goal is CLI-first, with MCP reserved for cases with no
CLI equivalent. The mesh-task lifecycle (the real, agent-claimable
`meshtask` model — as opposed to the legacy, disconnected `task` table
`edgeplane task create/list/show/update/delete` already covers) had no CLI
path at all: submit/claim/heartbeat/progress/complete/fail/block only
existed as `mcp__edgeplane__*` tools or raw REST. Adds `edgeplane task mesh
<verb>` mirroring all 9 tools 1:1, following the exact pattern
`WorkspaceCommand`/`handle_workspace` already establishes (build a JSON
args object, call `call_mcp_tool` -> POST /mcp/call — the same endpoint an
MCP client hits, just invoked as a CLI subprocess instead of a live tool
connection).

Two real bugs found while live-verifying the new commands against
production (`edgeplane task mesh progress` returned database_error):

- `progress_mesh_task` MCP handler: meshprogressevent INSERT never set
  `seq` (NOT NULL, no DB default) — same missing-column class as the
  submit_mesh_task/session_start bugs fixed in #110.
- Deeper, pre-existing bug in BOTH the new MCP handler and the REST
  `post_progress` handler (routes/work.rs, used by edgeplaned-work's task.rs
  client): the `SELECT COALESCE(MAX(seq),-1)+1` sequence query was decoded
  as `i64`, but `seq` is Postgres `integer` (i32) — sqlx's runtime decode
  silently failed on every call and fell through to unwrap_or(0). Every
  progress event ever posted, on either path, got seq=0 regardless of how
  many prior events existed for the task. Fixed the type in both files.

Regression tests (real Postgres): test_mcp_progress_mesh_task.rs and
test_work_progress_seq.rs both post 3 progress events on the same task and
assert seq comes back [0,1,2], not [0,0,0].

Verified: clippy --workspace --all-targets -D warnings clean; tower
199/199 (was 178/199 mid-branch before the seq fix); CLI 199/199 (+5 new);
workspace 499/499; launch-regression ok; no doc/dockerfile drift. Also
live-verified all 9 `edgeplane task mesh` subcommands against the prod
tower with a real domain/mission/task (submit/get/claim/heartbeat/
progress/complete/list all returned correct, non-404/500 responses).

+++ crates/edgeplane-tower/src/routes/work.rs
@@ -1155,8 +1155,10 @@ async fn append_progress(
         return resp;
     }
 
-    // Get next sequence number
-    let seq: i64 = sqlx::query_scalar(
+    // Get next sequence number. seq is `integer` (i32) in Postgres — decoding as i64
+    // here previously mismatched the wire type, so sqlx's runtime decode always failed
+    // and silently fell through to unwrap_or(0): every progress event got seq=0.
+    let seq: i32 = sqlx::query_scalar(
         "SELECT COALESCE(MAX(seq), -1) + 1 FROM meshprogressevent WHERE task_id=$1",
     )
     .bind(&task_id)
@@ -1177,7 +1179,7 @@ async fn append_progress(
     )
     .bind(&task_id)
     .bind(agent_id)
-    .bind(seq as i32)
+    .bind(seq)
     .bind(&body.event_type)
     .bind(&body.phase)
     .bind(&body.step)
